requirements.txt
playwright
playwright-stealth
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
chromadb
yt-dlp
python-docx
pymupdf
python-telegram-bot
schedule
python-dotenv
beautifulsoup4
requests
google-generativeai
langchain
langchain-community
langchain-google-genai
